Rename component folder and files to match naming convention. Add gap prop and gap class variations.
