+++ src/components/index.js
@@ -21,4 +21,4 @@ export { default as EditorInput } from './editor-input/index';
 export { default as ProgressSteps } from './progress-steps/index';
 export { default as Skeleton } from './skeleton/index';
 export { default as Menu } from './menu-item/index';
-export { default as Topbar } from './top-navbar/index';
+export { default as Topbar } from './topbar/index';
