--- src/components/top-navbar/index.js
@@ -1,1 +1,0 @@
-export { default } from './top-navbar.jsx';
